+++ axum/src/json.rs
@@ -226,7 +226,7 @@ mod tests {
     use serde::Deserialize;
     use serde_json::{json, Value};
 
-    #[tokio::test]
+    #[crate::test]
     async fn deserialize_body() {
         #[derive(Debug, Deserialize)]
         struct Input {
@@ -242,7 +242,7 @@ mod tests {
         assert_eq!(body, "bar");
     }
 
-    #[tokio::test]
+    #[crate::test]
     async fn consume_body_to_json_requires_json_content_type() {
         #[derive(Debug, Deserialize)]
         struct Input {
@@ -259,10 +259,10 @@ mod tests {
         assert_eq!(status, StatusCode::UNSUPPORTED_MEDIA_TYPE);
     }
 
-    #[tokio::test]
+    #[crate::test]
     async fn json_content_types() {
         async fn valid_json_content_type(content_type: &str) -> bool {
-            println!("testing {:?}", content_type);
+            println!("testing {content_type:?}");
 
             let app = Router::new().route("/", post(|Json(_): Json<Value>| async {}));
 
@@ -283,7 +283,7 @@ mod tests {
         assert!(!valid_json_content_type("text/json").await);
     }
 
-    #[tokio::test]
+    #[crate::test]
     async fn invalid_json_syntax() {
         let app = Router::new().route("/", post(|_: Json<serde_json::Value>| async {}));
 
@@ -314,7 +314,7 @@ mod tests {
         y: i32,
     }
 
-    #[tokio::test]
+    #[crate::test]
     async fn invalid_json_data() {
         let app = Router::new().route("/", post(|_: Json<Foo>| async {}));