+++ axum/src/routing/tests/mod.rs
@@ -29,7 +29,7 @@ mod handle_error;
 mod merge;
 mod nest;
 
-#[tokio::test]
+#[crate::test]
 async fn hello_world() {
     async fn root(_: Request<Body>) -> &'static str {
         "Hello, World!"
@@ -62,7 +62,7 @@ async fn hello_world() {
     assert_eq!(body, "users#create");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn routing() {
     let app = Router::new()
         .route(
@@ -98,7 +98,7 @@ async fn routing() {
     assert_eq!(res.text().await, "users#action");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn router_type_doesnt_change() {
     let app: Router = Router::new()
         .route(
@@ -123,7 +123,7 @@ async fn router_type_doesnt_change() {
     assert_eq!(res.text().await, "hi from POST");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn routing_between_services() {
     use std::convert::Infallible;
     use tower::service_fn;
@@ -169,7 +169,7 @@ async fn routing_between_services() {
     assert_eq!(res.text().await, "handler");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn middleware_on_single_route() {
     use tower::ServiceBuilder;
     use tower_http::{compression::CompressionLayer, trace::TraceLayer};
@@ -196,7 +196,7 @@ async fn middleware_on_single_route() {
     assert_eq!(body, "Hello, World!");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn service_in_bottom() {
     async fn handler(_req: Request<Body>) -> Result<Response<Body>, Infallible> {
         Ok(Response::new(hyper::Body::empty()))
@@ -207,7 +207,7 @@ async fn service_in_bottom() {
     TestClient::new(app);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn wrong_method_handler() {
     let app = Router::new()
         .route("/", get(|| async {}).post(|| async {}))
@@ -230,7 +230,7 @@ async fn wrong_method_handler() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn wrong_method_service() {
     #[derive(Clone)]
     struct Svc;
@@ -270,7 +270,7 @@ async fn wrong_method_service() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn multiple_methods_for_one_handler() {
     async fn root(_: Request<Body>) -> &'static str {
         "Hello, World!"
@@ -287,7 +287,7 @@ async fn multiple_methods_for_one_handler() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn wildcard_sees_whole_url() {
     let app = Router::new().route("/api/*rest", get(|uri: Uri| async move { uri.to_string() }));
 
@@ -297,7 +297,7 @@ async fn wildcard_sees_whole_url() {
     assert_eq!(res.text().await, "/api/foo/bar");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn middleware_applies_to_routes_above() {
     let app = Router::new()
         .route("/one", get(std::future::pending::<()>))
@@ -319,7 +319,7 @@ async fn middleware_applies_to_routes_above() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn not_found_for_extra_trailing_slash() {
     let app = Router::new().route("/foo", get(|| async {}));
 
@@ -332,7 +332,7 @@ async fn not_found_for_extra_trailing_slash() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn not_found_for_missing_trailing_slash() {
     let app = Router::new().route("/foo/", get(|| async {}));
 
@@ -342,7 +342,7 @@ async fn not_found_for_missing_trailing_slash() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn with_and_without_trailing_slash() {
     let app = Router::new()
         .route("/foo", get(|| async { "without tsr" }))
@@ -360,7 +360,7 @@ async fn with_and_without_trailing_slash() {
 }
 
 // for https://github.com/tokio-rs/axum/issues/420
-#[tokio::test]
+#[crate::test]
 async fn wildcard_doesnt_match_just_trailing_slash() {
     let app = Router::new().route(
         "/x/*path",
@@ -380,12 +380,12 @@ async fn wildcard_doesnt_match_just_trailing_slash() {
     assert_eq!(res.text().await, "foo/bar");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn static_and_dynamic_paths() {
     let app = Router::new()
         .route(
             "/:key",
-            get(|Path(key): Path<String>| async move { format!("dynamic: {}", key) }),
+            get(|Path(key): Path<String>| async move { format!("dynamic: {key}") }),
         )
         .route("/foo", get(|| async { "static" }));
 
@@ -398,14 +398,14 @@ async fn static_and_dynamic_paths() {
     assert_eq!(res.text().await, "static");
 }
 
-#[tokio::test]
+#[crate::test]
 #[should_panic(expected = "Paths must start with a `/`. Use \"/\" for root routes")]
 async fn empty_route() {
     let app = Router::new().route("", get(|| async {}));
     TestClient::new(app);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn middleware_still_run_for_unmatched_requests() {
     #[derive(Clone)]
     struct CountMiddleware<S>(S);
@@ -445,7 +445,7 @@ async fn middleware_still_run_for_unmatched_requests() {
     assert_eq!(COUNT.load(Ordering::SeqCst), 2);
 }
 
-#[tokio::test]
+#[crate::test]
 #[should_panic(expected = "\
     Invalid route: `Router::route_service` cannot be used with `Router`s. \
     Use `Router::nest` instead\
@@ -454,7 +454,7 @@ async fn routing_to_router_panics() {
     TestClient::new(Router::new().route_service("/", Router::new()));
 }
 
-#[tokio::test]
+#[crate::test]
 async fn route_layer() {
     let app = Router::new()
         .route("/foo", get(|| async {}))
@@ -482,7 +482,7 @@ async fn route_layer() {
     assert_eq!(res.status(), StatusCode::UNAUTHORIZED);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn different_methods_added_in_different_routes() {
     let app = Router::new()
         .route("/", get(|| async { "GET" }))
@@ -499,7 +499,7 @@ async fn different_methods_added_in_different_routes() {
     assert_eq!(body, "POST");
 }
 
-#[tokio::test]
+#[crate::test]
 #[should_panic(expected = "Cannot merge two `Router`s that both have a fallback")]
 async fn merging_routers_with_fallbacks_panics() {
     async fn fallback() {}
@@ -525,7 +525,7 @@ fn merging_with_overlapping_method_routes() {
     app.clone().merge(app);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn merging_routers_with_same_paths_but_different_methods() {
     let one = Router::new().route("/", get(|| async { "GET" }));
     let two = Router::new().route("/", post(|| async { "POST" }));
@@ -541,7 +541,7 @@ async fn merging_routers_with_same_paths_but_different_methods() {
     assert_eq!(body, "POST");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn head_content_length_through_hyper_server() {
     let app = Router::new()
         .route("/", get(|| async { "foo" }))
@@ -558,7 +558,7 @@ async fn head_content_length_through_hyper_server() {
     assert!(res.text().await.is_empty());
 }
 
-#[tokio::test]
+#[crate::test]
 async fn head_content_length_through_hyper_server_that_hits_fallback() {
     let app = Router::new().fallback(|| async { "foo" });
 
@@ -568,7 +568,7 @@ async fn head_content_length_through_hyper_server_that_hits_fallback() {
     assert_eq!(res.headers()["content-length"], "3");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn head_with_middleware_applied() {
     use tower_http::compression::{predicate::SizeAbove, CompressionLayer};
 
@@ -601,14 +601,14 @@ async fn head_with_middleware_applied() {
     assert!(!res.headers().contains_key("content-length"));
 }
 
-#[tokio::test]
+#[crate::test]
 #[should_panic(expected = "Paths must start with a `/`")]
 async fn routes_must_start_with_slash() {
     let app = Router::new().route(":foo", get(|| async {}));
     TestClient::new(app);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn body_limited_by_default() {
     let app = Router::new()
         .route("/bytes", post(|_: Bytes| async {}))
@@ -618,7 +618,7 @@ async fn body_limited_by_default() {
     let client = TestClient::new(app);
 
     for uri in ["/bytes", "/string", "/json"] {
-        println!("calling {}", uri);
+        println!("calling {uri}");
 
         let stream = futures_util::stream::repeat("a".repeat(1000)).map(Ok::<_, hyper::Error>);
         let body = Body::wrap_stream(stream);
@@ -636,7 +636,7 @@ async fn body_limited_by_default() {
     }
 }
 
-#[tokio::test]
+#[crate::test]
 async fn disabling_the_default_limit() {
     let app = Router::new()
         .route("/", post(|_: Bytes| async {}))
@@ -652,7 +652,7 @@ async fn disabling_the_default_limit() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn limited_body_with_content_length() {
     const LIMIT: usize = 3;
 
@@ -674,7 +674,7 @@ async fn limited_body_with_content_length() {
     assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn changing_the_default_limit() {
     let new_limit = 2;
 
@@ -699,7 +699,7 @@ async fn changing_the_default_limit() {
     assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn limited_body_with_streaming_body() {
     const LIMIT: usize = 3;
 
@@ -731,7 +731,7 @@ async fn limited_body_with_streaming_body() {
     assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn extract_state() {
     #[derive(Clone)]
     struct AppState {
@@ -767,7 +767,7 @@ async fn extract_state() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn explicitly_set_state() {
     let app = Router::new()
         .route_service(
@@ -781,7 +781,7 @@ async fn explicitly_set_state() {
     assert_eq!(res.text().await, "foo");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn layer_response_into_response() {
     fn map_response<B>(_res: Response<B>) -> Result<Response<B>, impl IntoResponse> {
         let headers = [("x-foo", "bar")];
@@ -798,4 +798,17 @@ async fn layer_response_into_response() {
     let res = client.get("/").send().await;
     assert_eq!(res.headers()["x-foo"], "bar");
     assert_eq!(res.status(), StatusCode::IM_A_TEAPOT);
+}
+
+#[allow(dead_code)]
+fn method_router_fallback_with_state() {
+    async fn fallback(_: State<&'static str>) {}
+
+    async fn not_found(_: State<&'static str>) {}
+
+    let state = "foo";
+
+    let _: Router = Router::new()
+        .fallback(get(fallback).fallback(not_found))
+        .with_state(state);
 }