+++ axum/src/routing/tests/nest.rs
@@ -3,7 +3,7 @@ use crate::{body::boxed, extract::Extension};
 use std::collections::HashMap;
 use tower_http::services::ServeDir;
 
-#[tokio::test]
+#[crate::test]
 async fn nesting_apps() {
     let api_routes = Router::new()
         .route(
@@ -58,7 +58,7 @@ async fn nesting_apps() {
     assert_eq!(res.text().await, "v0: games#show (123)");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn wrong_method_nest() {
     let nested_app = Router::new().route("/", get(|| async {}));
     let app = Router::new().nest("/", nested_app);
@@ -76,7 +76,7 @@ async fn wrong_method_nest() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nesting_router_at_root() {
     let nested = Router::new().route("/foo", get(|uri: Uri| async move { uri.to_string() }));
     let app = Router::new().nest("/", nested);
@@ -94,7 +94,7 @@ async fn nesting_router_at_root() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nesting_router_at_empty_path() {
     let nested = Router::new().route("/foo", get(|uri: Uri| async move { uri.to_string() }));
     let app = Router::new().nest("", nested);
@@ -112,7 +112,7 @@ async fn nesting_router_at_empty_path() {
     assert_eq!(res.status(), StatusCode::NOT_FOUND);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nesting_handler_at_root() {
     let app = Router::new().nest_service("/", get(|uri: Uri| async move { uri.to_string() }));
 
@@ -131,7 +131,7 @@ async fn nesting_handler_at_root() {
     assert_eq!(res.text().await, "/foo/bar");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nested_url_extractor() {
     let app = Router::new().nest(
         "/foo",
@@ -157,7 +157,7 @@ async fn nested_url_extractor() {
     assert_eq!(res.text().await, "/qux");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nested_url_original_extractor() {
     let app = Router::new().nest(
         "/foo",
@@ -177,7 +177,7 @@ async fn nested_url_original_extractor() {
     assert_eq!(res.text().await, "/foo/bar/baz");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nested_service_sees_stripped_uri() {
     let app = Router::new().nest(
         "/foo",
@@ -200,14 +200,14 @@ async fn nested_service_sees_stripped_uri() {
     assert_eq!(res.text().await, "/baz");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nest_static_file_server() {
     let app = Router::new().nest_service(
         "/static",
         get_service(ServeDir::new(".")).handle_error(|error| async move {
             (
                 StatusCode::INTERNAL_SERVER_ERROR,
-                format!("Unhandled internal error: {}", error),
+                format!("Unhandled internal error: {error}"),
             )
         }),
     );
@@ -218,7 +218,7 @@ async fn nest_static_file_server() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nested_multiple_routes() {
     let app = Router::new()
         .nest(
@@ -244,7 +244,7 @@ fn nested_at_root_with_other_routes() {
         .route("/", get(|| async {}));
 }
 
-#[tokio::test]
+#[crate::test]
 async fn multiple_top_level_nests() {
     let app = Router::new()
         .nest(
@@ -262,13 +262,13 @@ async fn multiple_top_level_nests() {
     assert_eq!(client.get("/two/route").send().await.text().await, "two");
 }
 
-#[tokio::test]
+#[crate::test]
 #[should_panic(expected = "Invalid route: nested routes cannot contain wildcards (*)")]
 async fn nest_cannot_contain_wildcards() {
     Router::<(), Body>::new().nest("/one/*rest", Router::new());
 }
 
-#[tokio::test]
+#[crate::test]
 async fn outer_middleware_still_see_whole_url() {
     #[derive(Clone)]
     struct SetUriExtension<S>(S);
@@ -319,11 +319,11 @@ async fn outer_middleware_still_see_whole_url() {
     assert_eq!(client.get("/one/two").send().await.text().await, "/one/two");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nest_at_capture() {
     let api_routes = Router::new().route(
         "/:b",
-        get(|Path((a, b)): Path<(String, String)>| async move { format!("a={} b={}", a, b) }),
+        get(|Path((a, b)): Path<(String, String)>| async move { format!("a={a} b={b}") }),
     );
 
     let app = Router::new().nest("/:a", api_routes);
@@ -335,7 +335,7 @@ async fn nest_at_capture() {
     assert_eq!(res.text().await, "a=foo b=bar");
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nest_with_and_without_trailing() {
     let app = Router::new().nest_service("/foo", get(|| async {}));
 
@@ -351,7 +351,7 @@ async fn nest_with_and_without_trailing() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn nesting_with_root_inner_router() {
     let app = Router::new().nest(
         "/foo",
@@ -371,7 +371,7 @@ async fn nesting_with_root_inner_router() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn fallback_on_inner() {
     let app = Router::new()
         .nest(
@@ -399,7 +399,7 @@ macro_rules! nested_route_test {
         // the route we expect to be able to call
         expected = $expected_path:literal $(,)?
     ) => {
-        #[tokio::test]
+        #[crate::test]
         async fn $name() {
             let inner = Router::new().route($route_path, get(|| async {}));
             let app = Router::new().nest($nested_path, inner);